Renamed config file and util.
