+++ recovery/cli/config/geo.ini
@@ -14,7 +14,7 @@ admin1 = stateProvince
 admin2 = county
 species = scientificName
 
-[matchnames]
+[matchfields]
 admin0 = ISO_A2
 admin1 = dptos
 admin2 = mpios
@@ -41,6 +41,14 @@ urban = urbanValidationSource
 [valuenames]
 elevation = extractedElevation
 
+[duplicates]
+gridres = 0.008333333767967150002
+defaultyear = none
+mark = all
+
 [misc]
-resolution = 0.008333333767967150002
-url = https://drive.google.com/uc?id=15TsDoq7u4hRotucE2LJyO-N5XCIteuCi
+crs = epsg:4326
+
+[behaviour]
+drop = False
+quiet = False
